+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 * Fix for ActiveRecord 5.2.4 (note security fix in 5.2.4.2 for ActiveView's escape_javascript CVE-2020-5267 for all earlier versions)
+* Fix issue where negative predicates with associated_collection does not use the correct primary key for subquery.
 
 ## 2.3.2 - 2020-01-11