+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+* Support ActiveRecord 6.1.0.rc1.
+  PR [1172](https://github.com/activerecord-hackery/ransack/pull/1172)
+
 * Fix for ActiveRecord 5.2.4 (note security fix in 5.2.4.2 for ActiveView's escape_javascript CVE-2020-5267 for all earlier versions)
 
 * Drop support for ActiveRecord older than 5.2.4.